stripped out unused tests, and merged api and index
