+++ views/layout.jade
@@ -13,10 +13,8 @@ html(ng-app='app')
     .container
       block content
 
-    //script(src="/components/jquery/dist/jquery.min.js")
     script(src='/components/angular/angular.min.js')
     script(src='/components/angular-bootstrap/ui-bootstrap.min.js')
 
-    //script(src="/components/bootstrap/dist/js/bootstrap.min.js")
     script(src='/socket.io/socket.io.js')
     script(src='/js/app.js')
